Move read_participants_file, read_file_sct, read_file_manual, and normalize_sensorimotor_scores functions into a new file, utils.py for reusability
